Update the extractor mixin to reduce code duplications
