+++ src/lib/colors/color-extractor.ts
@@ -7,14 +7,12 @@ export interface IColorStrategy extends IStrategy {
   extractColor(text: string): IColor;
 }
 class ColorExtractor extends Extractor {
-  public strategies: IColorStrategy[];
-
   public async extract(fileLines: DocumentLine[]): Promise < LineExtraction[] > {
-    const colors = await Promise.all(this.strategies.map(strategy => strategy.extractColors(fileLines)));
+    const colors = await Promise.all(this.enabledStrategies.map(strategy => (<IColorStrategy> strategy).extractColors(fileLines)));
     return flattenLineExtractionsFlatten(colors); // should regroup per lines?
   }
   public extractOneColor(text: string): IColor {
-    let colors = this.strategies.map(strategy => strategy.extractColor(text));
+    let colors = this.enabledStrategies.map(strategy => (<IColorStrategy> strategy).extractColor(text));
     return colors.find(color => color !== null);
   }
 }
